requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
PyJWT==2.9.0
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3